refactor: migrate to vite
