+++ package.json
@@ -1,7 +1,5 @@
 {
-  "name": "website-next",
-  "version": "0.1.0",
-  "private": true,
+  "type": "module",
   "dependencies": {
     "@typebot.io/js": "^0.2.41",
     "@typebot.io/nextjs": "^0.2.41",
@@ -13,13 +11,15 @@
     "react-terminal-component": "^1.5.0"
   },
   "devDependencies": {
+    "@tailwindcss/vite": "^4.1.13",
     "@types/node": "^20.0.0",
     "@types/react": "^18.0.0",
     "@types/react-dom": "^18.0.0",
+    "@vitejs/plugin-react": "^5.0.4",
     "autoprefixer": "^10.0.1",
     "postcss": "^8.0.0",
-    "tailwindcss": "^3.3.0",
-    "typescript": "^5.0.2"
+    "typescript": "^5.0.2",
+    "vite": "^7.1.7"
   },
   "engines": {
     "node": "22.x"
