--- src/components/Terminal/links.ts
@@ -1,7 +1,0 @@
-export const links = {
-  github: 'https://github.com/sripwoud',
-  repos: 'https://github.com/sripwoud?tab=repositories',
-  'cryptography-101': 'https://cryptography.sripwoud.xyz',
-  'zk-book-exercises': 'https://zk-book.sripwoud.xyz',
-  blog: 'https://listed.to/@sripwoud',
-}
