+++ src/const.ts
@@ -1,4 +1,4 @@
-export const FILE_SYSTEM = {
+export const fileSystem = {
   '/about': {},
   '/about/skills': {},
   '/about/skills/tech': {
@@ -52,4 +52,12 @@ French, English, German, (Spanish)
     content: 'Modern Cryptography Primitives 101.',
     canModify: false,
   },
+}
+
+export const links = {
+  github: 'https://github.com/sripwoud',
+  repos: 'https://github.com/sripwoud?tab=repositories',
+  'cryptography-101': 'https://cryptography.sripwoud.xyz',
+  'zk-book-exercises': 'https://zk-book.sripwoud.xyz',
+  blog: 'https://listed.to/@sripwoud',
 }
