+++ src/main.tsx
@@ -1,0 +1,11 @@
+import { App } from 'App'
+import { StrictMode } from 'react'
+import { createRoot } from 'react-dom/client'
+import './globals.css'
+
+// biome-ignore lint/style/noNonNullAssertion: index.html contains div#root
+createRoot(document.getElementById('root')!).render(
+  <StrictMode>
+    <App />
+  </StrictMode>,
+)
